Len changed from int to size_t in DNS

+++ src/analyzer/protocol/dns/DNS.cc
@@ -23,7 +23,7 @@ DNS_Interpreter::DNS_Interpreter(analyzer::Analyzer* arg_analyzer)
 	first_message = true;
 	}
 
-int DNS_Interpreter::ParseMessage(const u_char* data, int len, int is_query)
+int DNS_Interpreter::ParseMessage(const u_char* data, size_t len, int is_query)
 	{
 	int hdr_len = sizeof(DNS_RawMsgHdr);
 
@@ -143,7 +143,7 @@ int DNS_Interpreter::EndMessage(DNS_MsgInfo* msg)
 	}
 
 int DNS_Interpreter::ParseQuestions(DNS_MsgInfo* msg,
-				const u_char*& data, int& len,
+				const u_char*& data, size_t& len,
 				const u_char* msg_start)
 	{
 	int n = msg->qdcount;
@@ -154,7 +154,7 @@ int DNS_Interpreter::ParseQuestions(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseAnswers(DNS_MsgInfo* msg, int n, DNS_AnswerType atype,
-				const u_char*& data, int& len,
+				const u_char*& data, size_t& len,
 				const u_char* msg_start)
 	{
 	msg->answer_type = atype;
@@ -166,11 +166,11 @@ int DNS_Interpreter::ParseAnswers(DNS_MsgInfo* msg, int n, DNS_AnswerType atype,
 	}
 
 int DNS_Interpreter::ParseQuestion(DNS_MsgInfo* msg,
-				const u_char*& data, int& len,
+				const u_char*& data, size_t& len,
 				const u_char* msg_start)
 	{
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 	if ( ! name_end )
@@ -212,11 +212,11 @@ int DNS_Interpreter::ParseQuestion(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseAnswer(DNS_MsgInfo* msg,
-				const u_char*& data, int& len,
+				const u_char*& data, size_t& len,
 				const u_char* msg_start)
 	{
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 
@@ -357,8 +357,8 @@ int DNS_Interpreter::ParseAnswer(DNS_MsgInfo* msg,
 	return status;
 	}
 
-u_char* DNS_Interpreter::ExtractName(const u_char*& data, int& len,
-					u_char* name, int name_len,
+u_char* DNS_Interpreter::ExtractName(const u_char*& data, size_t& len,
+					u_char* name, size_t name_len,
 					const u_char* msg_start)
 	{
 	u_char* name_start = name;
@@ -366,7 +366,7 @@ u_char* DNS_Interpreter::ExtractName(const u_char*& data, int& len,
 	while ( ExtractLabel(data, len, name, name_len, msg_start) )
 		;
 
-	int n = name - name_start;
+	size_t n = name - name_start;
 
 	if ( n >= 255 )
 		analyzer->Weird("DNS_NAME_too_long");
@@ -386,8 +386,8 @@ u_char* DNS_Interpreter::ExtractName(const u_char*& data, int& len,
 	return name;
 	}
 
-int DNS_Interpreter::ExtractLabel(const u_char*& data, int& len,
-				u_char*& name, int& name_len,
+int DNS_Interpreter::ExtractLabel(const u_char*& data, size_t& len,
+				u_char*& name, size_t& name_len,
 				const u_char* msg_start)
 	{
 	if ( len <= 0 )
@@ -431,7 +431,7 @@ int DNS_Interpreter::ExtractLabel(const u_char*& data, int& len,
 
 		// Recursively resolve name.
 		const u_char* recurse_data = msg_start + offset;
-		int recurse_max_len = orig_data - recurse_data;
+		size_t recurse_max_len = orig_data - recurse_data;
 
 		u_char* name_end = ExtractName(recurse_data, recurse_max_len,
 						name, name_len, msg_start);
@@ -476,7 +476,7 @@ int DNS_Interpreter::ExtractLabel(const u_char*& data, int& len,
 	return 1;
 	}
 
-uint16_t DNS_Interpreter::ExtractShort(const u_char*& data, int& len)
+uint16_t DNS_Interpreter::ExtractShort(const u_char*& data, size_t& len)
 	{
 	if ( len < 2 )
 		return 0;
@@ -496,7 +496,7 @@ uint16_t DNS_Interpreter::ExtractShort(const u_char*& data, int& len)
 	return val;
 	}
 
-uint32_t DNS_Interpreter::ExtractLong(const u_char*& data, int& len)
+uint32_t DNS_Interpreter::ExtractLong(const u_char*& data, size_t& len)
 	{
 	if ( len < 4 )
 		return 0;
@@ -515,13 +515,13 @@ uint32_t DNS_Interpreter::ExtractLong(const u_char*& data, int& len)
 	}
 
 int DNS_Interpreter::ParseRR_Name(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	const u_char* data_start = data;
 
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 	if ( ! name_end )
@@ -567,20 +567,20 @@ int DNS_Interpreter::ParseRR_Name(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_SOA(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	const u_char* data_start = data;
 
 	u_char mname[513];
-	int mname_len = sizeof(mname) - 1;
+	size_t mname_len = sizeof(mname) - 1;
 
 	u_char* mname_end = ExtractName(data, len, mname, mname_len, msg_start);
 	if ( ! mname_end )
 		return 0;
 
 	u_char rname[513];
-	int rname_len = sizeof(rname) - 1;
+	size_t rname_len = sizeof(rname) - 1;
 
 	u_char* rname_end = ExtractName(data, len, rname, rname_len, msg_start);
 	if ( ! rname_end )
@@ -621,7 +621,7 @@ int DNS_Interpreter::ParseRR_SOA(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_MX(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	const u_char* data_start = data;
@@ -629,7 +629,7 @@ int DNS_Interpreter::ParseRR_MX(DNS_MsgInfo* msg,
 	int preference = ExtractShort(data, len);
 
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 	if ( ! name_end )
@@ -653,7 +653,7 @@ int DNS_Interpreter::ParseRR_MX(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_NBS(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	data += rdlength;
@@ -662,7 +662,7 @@ int DNS_Interpreter::ParseRR_NBS(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_SRV(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	const u_char* data_start = data;
@@ -672,7 +672,7 @@ int DNS_Interpreter::ParseRR_SRV(DNS_MsgInfo* msg,
 	unsigned int port = ExtractShort(data, len);
 
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 	if ( ! name_end )
@@ -698,7 +698,7 @@ int DNS_Interpreter::ParseRR_SRV(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_EDNS(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	// We need a pair-value set mechanism here to dump useful information
@@ -725,11 +725,11 @@ int DNS_Interpreter::ParseRR_EDNS(DNS_MsgInfo* msg,
 	return 1;
 	}
 
-void DNS_Interpreter::ExtractOctets(const u_char*& data, int& len,
+void DNS_Interpreter::ExtractOctets(const u_char*& data, size_t& len,
                                     BroString** p)
 	{
 	uint16_t dlen = ExtractShort(data, len);
-	dlen = min(len, static_cast<int>(dlen));
+	dlen = min(static_cast<int>(len), static_cast<int>(dlen));
 
 	if ( p )
 		*p = new BroString(data, dlen, 0);
@@ -738,10 +738,10 @@ void DNS_Interpreter::ExtractOctets(const u_char*& data, int& len,
 	len -= dlen;
 	}
 
-BroString* DNS_Interpreter::ExtractStream(const u_char*& data, int& len, int l)
+BroString* DNS_Interpreter::ExtractStream(const u_char*& data, size_t& len, int l)
 	{
 	l = max(l, 0);
-	int dlen = min(len, l); // Len in bytes of the algorithm use
+	size_t dlen = min(len, static_cast<size_t>(l)); // Len in bytes of the algorithm use
 	auto rval = new BroString(data, dlen, 0);
 
 	data += dlen;
@@ -750,12 +750,12 @@ BroString* DNS_Interpreter::ExtractStream(const u_char*& data, int& len, int l)
 	}
 
 int DNS_Interpreter::ParseRR_TSIG(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	const u_char* data_start = data;
 	u_char alg_name[1024];
-	int alg_name_len = sizeof(alg_name) - 1;
+	size_t alg_name_len = sizeof(alg_name) - 1;
 
 	u_char* alg_name_end =
 		ExtractName(data, len, alg_name, alg_name_len, msg_start);
@@ -795,7 +795,7 @@ int DNS_Interpreter::ParseRR_TSIG(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_RRSIG(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_RRSIG || msg->skip_event )
@@ -822,7 +822,7 @@ int DNS_Interpreter::ParseRR_RRSIG(DNS_MsgInfo* msg,
 	//implement signer's name with the msg_start offset
 	const u_char* data_start = data;
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 	if ( ! name_end )
@@ -897,7 +897,7 @@ int DNS_Interpreter::ParseRR_RRSIG(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_DNSKEY(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_DNSKEY || msg->skip_event )
@@ -992,7 +992,7 @@ int DNS_Interpreter::ParseRR_DNSKEY(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_NSEC(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_NSEC || msg->skip_event )
@@ -1004,7 +1004,7 @@ int DNS_Interpreter::ParseRR_NSEC(DNS_MsgInfo* msg,
 
 	const u_char* data_start = data;
 	u_char name[513];
-	int name_len = sizeof(name) - 1;
+	size_t name_len = sizeof(name) - 1;
 
 	u_char* name_end = ExtractName(data, len, name, name_len, msg_start);
 	if ( ! name_end )
@@ -1046,7 +1046,7 @@ int DNS_Interpreter::ParseRR_NSEC(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_NSEC3(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_NSEC3 || msg->skip_event )
@@ -1132,7 +1132,7 @@ int DNS_Interpreter::ParseRR_NSEC3(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_DS(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_DS || msg->skip_event )
@@ -1190,7 +1190,7 @@ int DNS_Interpreter::ParseRR_DS(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_A(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength)
+				const u_char*& data, size_t& len, int rdlength)
 	{
 	if ( rdlength != 4 )
 		{
@@ -1214,7 +1214,7 @@ int DNS_Interpreter::ParseRR_A(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_AAAA(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength)
+				const u_char*& data, size_t& len, int rdlength)
 	{
 	uint32_t addr[4];
 
@@ -1251,7 +1251,7 @@ int DNS_Interpreter::ParseRR_AAAA(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_WKS(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength)
+				const u_char*& data, size_t& len, int rdlength)
 	{
 	data += rdlength;
 	len -= rdlength;
@@ -1260,7 +1260,7 @@ int DNS_Interpreter::ParseRR_WKS(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_HINFO(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength)
+				const u_char*& data, size_t& len, int rdlength)
 	{
 	data += rdlength;
 	len -= rdlength;
@@ -1269,7 +1269,7 @@ int DNS_Interpreter::ParseRR_HINFO(DNS_MsgInfo* msg,
 	}
 
 static StringVal* extract_char_string(analyzer::Analyzer* analyzer,
-                                      const u_char*& data, int& len, int& rdlen)
+                                      const u_char*& data, size_t& len, int& rdlen)
 	{
 	if ( rdlen <= 0 )
 		return 0;
@@ -1297,7 +1297,7 @@ static StringVal* extract_char_string(analyzer::Analyzer* analyzer,
 	}
 
 int DNS_Interpreter::ParseRR_TXT(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_TXT_reply || msg->skip_event )
@@ -1327,7 +1327,7 @@ int DNS_Interpreter::ParseRR_TXT(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_SPF(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_SPF_reply || msg->skip_event )
@@ -1357,7 +1357,7 @@ int DNS_Interpreter::ParseRR_SPF(DNS_MsgInfo* msg,
 	}
 
 int DNS_Interpreter::ParseRR_CAA(DNS_MsgInfo* msg,
-				const u_char*& data, int& len, int rdlength,
+				const u_char*& data, size_t& len, int rdlength,
 				const u_char* msg_start)
 	{
 	if ( ! dns_CAA_reply || msg->skip_event )
@@ -1407,7 +1407,7 @@ int DNS_Interpreter::ParseRR_CAA(DNS_MsgInfo* msg,
 
 void DNS_Interpreter::SendReplyOrRejectEvent(DNS_MsgInfo* msg,
 						EventHandlerPtr event,
-						const u_char*& data, int& len,
+						const u_char*& data, size_t& len,
 						BroString* question_name)
 	{
 	RR_Type qtype = RR_Type(ExtractShort(data, len));
